Move `testBoth` and `testWithDefault` to `internal-test-helpers`.
