+++ packages/ember-metal/tests/accessors/get_test.js
@@ -1,4 +1,4 @@
-import { testBoth } from '../props_helper';
+import { testBoth } from 'internal-test-helpers';
 import {
   get,
   getWithDefault
