+++ packages/ember-metal/tests/binding/connect_test.js
@@ -1,5 +1,5 @@
 import { context } from 'ember-environment';
-import { testBoth } from '../props_helper';
+import { testBoth } from 'internal-test-helpers';
 import {
   Binding,
   bind
