+++ packages/ember-metal/tests/computed_test.js
@@ -1,5 +1,5 @@
 import { Object as EmberObject } from 'ember-runtime';
-import { testBoth } from './props_helper';
+import { testBoth } from 'internal-test-helpers';
 import {
   ComputedProperty,
   computed,
