+++ packages/ember-metal/tests/observer_test.js
@@ -1,5 +1,5 @@
 import { ENV } from 'ember-environment';
-import { testBoth } from './props_helper';
+import { testBoth } from 'internal-test-helpers';
 import {
   addObserver,
   removeObserver,
